testing

+++ test_neon2.c
@@ -127,8 +127,8 @@ void main(){
     //Prints the augmented matrix
     printMatrix( I, n, i, j );
 
-    int16_t a = 6570;
-    int16_t b = 540 * 100;
+    int32_t a = 6570;
+    int32_t b = 540 * 100;
     int32_t c = (b / a);
     printf("%i \n\n", a);
     printf("%i \n\n", b);
